s3: move stream logic to separate file, implement ReaderFrom
