+++ backends/s3/s3.go
@@ -462,73 +462,6 @@ func New(ctx context.Context, opt Options) (*Backend, error) {
 	return b, nil
 }
 
-// NewReader satisfies StreamReader and provides a read streaming interface to
-// a blob located on an S3 server.
-func (b *Backend) NewReader(ctx context.Context, name string) (io.ReadCloser, error) {
-	name = b.prependGlobalPrefix(name)
-	r, err := b.doLoadReader(ctx, name)
-	if err != nil {
-		return nil, err
-	}
-	return r, nil
-}
-
-// NewWriter satisfies StreamWriter and provides a write streaming interface to
-// a blob located on an S3 server.
-func (b *Backend) NewWriter(ctx context.Context, name string) (io.WriteCloser, error) {
-	name = b.prependGlobalPrefix(name)
-	pr, pw := io.Pipe()
-	wrap := &writerWrapper{
-		ctx:     ctx,
-		backend: b,
-		name:    name,
-		pw:      pw,
-		done:    make(chan struct{}, 1),
-	}
-	go func() {
-		var err error
-		wrap.info, err = b.doStoreReader(ctx, name, pr, -1)
-		_ = pr.CloseWithError(err)
-		wrap.done <- struct{}{}
-	}()
-	return wrap, nil
-}
-
-// A writerWrapper implements io.WriteCloser and is returned by (*Backend).NewWriter.
-type writerWrapper struct {
-	backend *Backend
-	done    chan struct{}
-
-	// We need to keep these around
-	// to write the marker in Close.
-	ctx  context.Context
-	info minio.UploadInfo
-	name string
-
-	// Writes are sent to this pipe
-	// and then written to S3 in a background goroutine.
-	pw *io.PipeWriter
-}
-
-func (w *writerWrapper) Write(p []byte) (int, error) {
-	return w.pw.Write(p)
-}
-
-func (w *writerWrapper) Close() error {
-	err := w.pw.Close()
-	if err != nil {
-		return err
-	}
-	select {
-	case <-w.done:
-	case <-w.ctx.Done():
-		err := w.ctx.Err()
-		_ = w.pw.CloseWithError(err)
-		return err
-	}
-	return w.backend.setMarker(w.ctx, w.name, w.info.ETag, false)
-}
-
 // setGlobalPrefix updates the global prefix in b and the cached marker name,
 // so it can be dynamically changed in tests.
 func (b *Backend) setGlobalPrefix(prefix string) {
